fix(server): :bug: fix errors on client disconnect
Two issues were occuring on client disconnect that have been fixed: (1) the disconnect handler would run even when who was not in a room left, (2) both disconnect handlers (senderLeave and receiverLeave) would run for both senders and receivers. These have been mitigated by explicitly checking if roomID is not defined and adding a clientType attribute to the socket instance.